extern: codis-server 3.2.7, using maxbytes to limit pipeline size

+++ extern/redis-3.2.7/src/server.h
@@ -727,7 +727,7 @@ typedef struct {
     int port;
     long long timeout;
     long long lastuse;
-    int pending_msgs;
+    long pending_msgs;
     void *batched_iter;
     list *blocked_list;
 } slotsmgrtAsyncClient;
